# swagger.yaml
openapi: 3.0.3
...
info:
  title: Laundry Service API
  description: |
    ## 🧺 Tổng quan
    API cho nền tảng dịch vụ Giặt ủi trực tuyến. Hệ thống hỗ trợ:
    - **Khách hàng (Customer)**: Đặt đơn giặt, theo dõi trạng thái, thanh toán
    - **Nhân viên (Staff)**: Xử lý đơn hàng, cập nhật trạng thái, upload ảnh
    - **Quản trị viên (Admin)**: Toàn quyền quản lý hệ thống

    ## 🔐 Xác thực
    Hệ thống hỗ trợ **2 luồng đăng nhập**:

    ### Luồng 1: Email + Password (dành cho Admin & Staff)
    ```
    POST /auth/register → POST /auth/login → nhận { accessToken }
    ```

    ### Luồng 2: SĐT + OTP (dành cho Customer — không cần password)
    ```
    POST /auth/send-otp → POST /auth/verify-otp → nhận { accessToken }
    ```

    ### Cách truyền Token lên Header
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ```

    ### Refresh Token
    - Được lưu trong **HttpOnly Cookie** (`refreshToken`), browser tự gửi kèm
    - Khi access token hết hạn, gọi `POST /auth/refresh-token` để cấp lại

    ## 📦 Luồng trạng thái đơn hàng
    ```
    received → washing → drying → delivering → completed
        ↓          ↓        ↓           ↓
    cancelled  cancelled cancelled  cancelled
    ```
    *(Customer chỉ hủy được khi status = received. Admin hủy được mọi lúc trừ completed)*

    ## 📝 Cấu trúc Response chuẩn
    Mọi response đều theo format:
    ```json
    {
      "success": true,
      "message": "Mô tả kết quả",
      "data": { ... },
...
    }
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token. Lấy từ response của `/auth/login` hoặc `/auth/verify-otp`.
        Truyền vào header: `Authorization: Bearer <token>`

  # ═══════════════════════════════════════════════════════════════════════════
  #  SCHEMAS (Data Models / Types)
  #  FE có thể dùng phần này để tạo TypeScript Interface/Type
  # ═══════════════════════════════════════════════════════════════════════════
  schemas:
...
    StaffNote:
      type: object
      description: Ghi chú của nhân viên về đơn hàng
      properties:
        _id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d9"
        content:
          type: string
          example: "Áo sơ mi bị ố vàng nhẹ, đã dùng tẩy chuyên dụng"
          maxLength: 500
        createdBy:
          oneOf:
            - $ref: '#/components/schemas/UserRef'
            - type: string
          description: Nhân viên tạo ghi chú
        createdAt:
          type: string
          format: date-time
          example: "2024-06-20T10:00:00.000Z"

    # ─── USER REF (Thông tin user tóm tắt khi populate) ─────────────────────
...
    Order:
      type: object
      description: Đơn hàng giặt ủi
      properties:
        _id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d3"
        orderCode:
          type: string
          example: "LD-20240620-0001"
          description: Mã đơn hàng tự động sinh (format LD-YYYYMMDD-XXXX)
        customer:
          oneOf:
            - $ref: '#/components/schemas/UserRef'
            - type: string
          description: Khách hàng (ObjectId hoặc object khi populate)
        staff:
          oneOf:
            - $ref: '#/components/schemas/UserRef'
            - type: string
          description: Nhân viên được phân công (null nếu chưa assign)
          nullable: true
        service:
          oneOf:
            - $ref: '#/components/schemas/ServiceRef'
            - type: string
          description: Gói dịch vụ (ObjectId hoặc object khi populate)
        quantity:
          type: number
          example: 3.5
          description: Số lượng (kg nếu priceType=per_kg, hoặc số món nếu per_item)
          minimum: 0.1
        totalPrice:
          type: number
          example: 87500
          description: Tổng tiền (VNĐ) = service.price × quantity, tính tự động
        status:
          type: string
          enum: [received, washing, drying, delivering, completed, cancelled]
          example: "received"
          description: |
            Trạng thái đơn hàng:
            - `received` (📦 Đã nhận đơn): Đơn mới tạo, chờ nhân viên lấy
            - `washing` (🫧 Đang giặt): Đang trong máy giặt
            - `drying` (🌬️ Đang sấy/ủi): Đang sấy hoặc ủi
            - `delivering` (🚚 Đang giao): Shipper đang trên đường giao
            - `completed` (✅ Hoàn thành): Đã giao thành công
            - `cancelled` (❌ Đã hủy): Đơn bị hủy
        pickupAddress:
          type: string
          example: "123 Lê Lợi, Quận 1, TP.HCM"
          description: Địa chỉ lấy đồ
        deliveryAddress:
          type: string
          example: "456 Nguyễn Trãi, Quận 5, TP.HCM"
          description: Địa chỉ giao đồ (có thể khác địa chỉ lấy)
        scheduledPickupTime:
          type: string
          format: date-time
          example: "2024-06-21T09:00:00.000Z"
          description: Thời gian lấy đồ mong muốn của khách
          nullable: true
        note:
          type: string
          example: "Quần áo màu tối, giặt riêng. Tránh nhiệt độ cao."
          description: Ghi chú của khách khi đặt đơn
          nullable: true
        staffNotes:
          type: array
          description: Danh sách ghi chú của nhân viên (có thể nhiều, theo thời gian)
          items:
            $ref: '#/components/schemas/StaffNote'
        adminNote:
          type: string
          example: "VIP customer — ưu tiên xử lý"
          description: Ghi chú riêng của Admin (1 ghi chú, ghi đè được)
          nullable: true
        statusHistory:
          type: array
          description: Lịch sử thay đổi trạng thái (audit trail)
          items:
            $ref: '#/components/schemas/StatusHistoryItem'
        completedAt:
          type: string
          format: date-time
          example: "2024-06-22T14:30:00.000Z"
          description: Thời điểm hoàn thành (tự động set khi status = completed)
          nullable: true
        createdAt:
          type: string
          format: date-time
          example: "2024-06-20T08:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-06-21T10:00:00.000Z"

    # ─── ORDER IMAGE ──────────────────────────────────────────────────────────
...
    CreateOrderRequest:
      type: object
      required: [serviceId, quantity, pickupAddress, deliveryAddress]
      description: Body để tạo đơn hàng mới
      properties:
        serviceId:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d2"
          description: ID của gói dịch vụ (lấy từ GET /services)
        quantity:
          type: number
          example: 3.5
          minimum: 0.1
          description: Số lượng kg (nếu priceType=per_kg) hoặc số món (nếu per_item)
        pickupAddress:
          type: string
          example: "123 Lê Lợi, Quận 1, TP.HCM"
          description: Địa chỉ nhân viên đến lấy đồ
        deliveryAddress:
          type: string
          example: "456 Nguyễn Trãi, Quận 5, TP.HCM"
          description: Địa chỉ giao đồ lại (có thể giống hoặc khác pickupAddress)
        scheduledPickupTime:
          type: string
          format: date-time
          example: "2024-06-21T09:00:00.000Z"
          description: Thời gian lấy đồ mong muốn (optional, mặc định sớm nhất có thể)
          nullable: true
        note:
          type: string
          example: "Quần áo màu tối, giặt riêng. Không giặt với đồ trắng."
          description: Ghi chú đặc biệt cho đơn hàng
          nullable: true
    UpdateOrderStatusRequest:
      type: object
      required: [status]
      description: Body để cập nhật trạng thái đơn hàng (dùng cho cả Admin và Staff routes)
      properties:
        status:
          type: string
          enum: [received, washing, drying, delivering, completed, cancelled]
          example: "washing"
          description: |
            Trạng thái mới. Phải tuân theo luồng hợp lệ:
            - received → washing | cancelled
            - washing → drying | cancelled
            - drying → delivering | cancelled
            - delivering → completed | cancelled
            - completed → (không thể chuyển)
            - cancelled → (không thể phục hồi)
        note:
          type: string
          example: "Bắt đầu giặt lúc 9h sáng"
          description: Ghi chú kèm theo khi chuyển trạng thái (optional)
          nullable: true
...
    AddStaffNoteRequest:
      type: object
      required: [content]
      description: Body để thêm ghi chú nhân viên
      properties:
        content:
          type: string
          example: "Áo sơ mi bị ố vàng nhẹ, đã dùng tẩy chuyên dụng"
          maxLength: 500
    UpdateAdminNoteRequest:
      type: object
      required: [adminNote]
      description: Body để cập nhật ghi chú admin (overwrite)
      properties:
        adminNote:
          type: string
          example: "VIP customer — ưu tiên xử lý"
          description: Ghi chú mới (truyền chuỗi rỗng "" để xóa ghi chú)
          maxLength: 300
    AssignStaffRequest:
      type: object
      required: [staffId]
      description: Body để phân công nhân viên cho đơn hàng
      properties:
        staffId:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d5"
          description: ID của nhân viên (role = staff)

    # ─── RESPONSE WRAPPERS ─────────────────────────────────────────────────────
...
    OrderListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Lấy danh sách đơn hàng thành công"
        data:
          type: object
          properties:
            orders:
              type: array
              items:
                $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/Pagination'
    DashboardResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Lấy thống kê dashboard thành công"
        data:
          $ref: '#/components/schemas/DashboardStats'

    # ─── ERROR RESPONSES ───────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      description: Cấu trúc lỗi chung
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Mô tả lỗi cụ thể"
        data:
          nullable: true
          example: null
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: MongoDB ObjectId của bản ghi
      schema:
        type: string
        example: "64f1a2b3c4d5e6f7a8b9c0d3"
    PageParam:
      name: page
      in: query
      description: Số trang (bắt đầu từ 1)
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      description: Số bản ghi mỗi trang
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100

  # ─── REUSABLE RESPONSES ───────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Token không hợp lệ hoặc chưa đăng nhập
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Bạn chưa đăng nhập. Vui lòng đăng nhập để tiếp tục."
            data: null
    Forbidden:
      description: Không có quyền truy cập
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Bạn không có quyền thực hiện hành động này."
            data: null
    NotFound:
      description: Không tìm thấy bản ghi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Không tìm thấy bản ghi"
            data: null
    BadRequest:
      description: Dữ liệu đầu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
security:
  - BearerAuth: []

# ═══════════════════════════════════════════════════════════════════════════════
#  TAGS — Nhóm các endpoint
# ═══════════════════════════════════════════════════════════════════════════════
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: Lấy danh sách đơn hàng (phân quyền theo role)
      description: |
        **Phân quyền theo role:**
        - **Customer:** Chỉ thấy đơn hàng của chính mình
        - **Staff & Admin:** Thấy tất cả đơn hàng, có thể lọc theo status

        Kết quả được phân trang và sắp xếp theo thời gian tạo (mới nhất trước).
      operationId: getAllOrders
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Lọc theo trạng thái đơn hàng
          schema:
            type: string
            enum: [received, washing, drying, delivering, completed, cancelled]
        - $ref: '#/components/parameters/PageParam'
        - name: limit
          in: query
          description: Số bản ghi mỗi trang
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lấy danh sách thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Orders]
      summary: Tạo đơn hàng mới
      description: |
        **[Yêu cầu đăng nhập — Customer]**

        Tạo đơn hàng giặt ủi mới. Tổng tiền được tính tự động:
        ```
        totalPrice = service.price × quantity
        ```

        Sau khi tạo thành công:
        - Đơn hàng có trạng thái ban đầu là `received`
        - Mã đơn hàng được tự động sinh (format: `LD-YYYYMMDD-XXXX`)
        - Staff dashboard nhận thông báo real-time qua Socket.io
      operationId: createOrder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              serviceId: "64f1a2b3c4d5e6f7a8b9c0d2"
              quantity: 3.5
              pickupAddress: "123 Lê Lợi, Quận 1, TP.HCM"
              deliveryAddress: "456 Nguyễn Trãi, Quận 5, TP.HCM"
              scheduledPickupTime: "2024-06-21T09:00:00.000Z"
              note: "Quần áo màu tối, giặt riêng. Tránh nhiệt cao."
      responses:
        '201':
          description: Tạo đơn hàng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Tạo đơn hàng thành công"
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'
        '400':
          description: Dữ liệu không hợp lệ hoặc dịch vụ không tồn tại/đã ngừng hoạt động
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /orders/{id}/status:
    put:
      tags: [Orders]
      summary: "[Staff/Admin] Cập nhật trạng thái đơn hàng"
      description: |
        **[Yêu cầu role: staff hoặc admin]**

        Cập nhật trạng thái đơn hàng theo đúng luồng:
        ```
        received → washing → drying → delivering → completed
        ```
        Bất kỳ trạng thái nào cũng có thể chuyển sang `cancelled` (trừ `completed`).

        > ⚠️ Nếu chuyển sang trạng thái không hợp lệ, server trả về lỗi 400 với
        > danh sách trạng thái hợp lệ tiếp theo.
      operationId: updateOrderStatus
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
            example:
              status: "washing"
              note: "Bắt đầu giặt lúc 9h sáng, máy giặt số 3"
      responses:
        '200':
          description: Cập nhật trạng thái thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Cập nhật trạng thái đơn hàng thành công → \"washing\""
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'
        '400':
          description: Trạng thái không hợp lệ hoặc vi phạm luồng trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Không thể chuyển từ trạng thái \"completed\" sang \"washing\". Các trạng thái hợp lệ tiếp theo: không có"
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /orders/{id}/assign:
    put:
      tags: [Admin — Staff Management]
      summary: "[Admin] Phân công nhân viên cho đơn hàng"
      description: |
        **[Yêu cầu role: admin]**

        Gán một nhân viên (staff) cho đơn hàng.
        Dùng để điều phối công việc trong nội bộ.
      operationId: assignStaff
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignStaffRequest'
            example:
              staffId: "64f1a2b3c4d5e6f7a8b9c0d5"
      responses:
        '200':
          description: Phân công thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Phân công nhân viên thành công"
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────────────────────────────────────
  #  PAYMENTS ENDPOINTS  (Base: /api/payments)
  # ─────────────────────────────────────────────────────────────────────────────
...
  /admin/dashboard:
    get:
      tags: [Admin — Dashboard & Orders]
      summary: "[Admin/Staff] Xem thống kê Dashboard"
      description: |
        **[Yêu cầu role: admin hoặc staff]**

        Trả về thống kê tổng quan cho trang Admin Dashboard:
        - Số đơn và doanh thu trong ngày hôm nay
        - Số đơn theo từng trạng thái
        - Doanh thu theo từng tháng trong năm
        - 10 đơn hàng gần nhất
      operationId: getDashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lấy thống kê thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/orders/{id}/staff-notes:
    post:
      tags: [Admin — Dashboard & Orders]
      summary: "[Admin/Staff] Thêm ghi chú nhân viên vào đơn hàng"
      description: |
        **[Yêu cầu role: admin hoặc staff]**

        Thêm một ghi chú mới vào danh sách `staffNotes` của đơn hàng.
        Ghi chú được **append** (không ghi đè ghi chú cũ).

        Ví dụ sử dụng:
        - "Áo sơ mi bị ố vàng nhẹ, đã dùng tẩy chuyên dụng"
        - "Khách hẹn giao sau 5h chiều"
        - "Quần jean màu đậm, giặt riêng"
      operationId: addStaffNote
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddStaffNoteRequest'
            example:
              content: "Áo sơ mi bị ố vàng nhẹ, đã dùng tẩy chuyên dụng. Kết quả tốt."
      responses:
        '201':
          description: Thêm ghi chú thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Thêm ghi chú thành công"
                  data:
                    type: object
                    properties:
                      staffNotes:
                        type: array
                        items:
                          $ref: '#/components/schemas/StaffNote'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/orders/{id}/admin-note:
    put:
      tags: [Admin — Dashboard & Orders]
      summary: "[Admin Only] Cập nhật ghi chú Admin"
      description: |
        **[Yêu cầu role: admin — Staff không được phép]**

        Cập nhật ghi chú nội bộ của Admin cho đơn hàng (overwrite — chỉ 1 ghi chú tại 1 thời điểm).
        Truyền chuỗi rỗng `""` để xóa ghi chú.

        Ví dụ sử dụng:
        - "VIP customer — ưu tiên xử lý"
        - "Đã xác nhận qua điện thoại lúc 10h sáng"
        - "Miễn phí ship do lần đầu tiên dùng dịch vụ"
      operationId: updateAdminNote
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminNoteRequest'
            examples:
              set_note:
                summary: Đặt ghi chú mới
                value:
                  adminNote: "VIP customer — ưu tiên xử lý"
              clear_note:
                summary: Xóa ghi chú (truyền chuỗi rỗng)
                value:
                  adminNote: ""
      responses:
        '200':
          description: Cập nhật ghi chú thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Cập nhật ghi chú admin thành công"
                  data:
                    type: object
                    properties:
                      adminNote:
                        type: string
                        example: "VIP customer — ưu tiên xử lý"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/staff:
    get:
      tags: [Admin — Staff Management]
      summary: "[Admin Only] Lấy danh sách nhân viên"
      description: |
        **[Yêu cầu role: admin — Staff không được phép xem]**

        Lấy danh sách nhân viên đang hoạt động (`isActive = true`, `role = staff`).
        Dùng để hiển thị dropdown chọn nhân viên khi phân công đơn hàng.
      operationId: getStaffList
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          description: Tìm kiếm theo tên, email hoặc số điện thoại
          schema:
            type: string
            example: "Nguyễn"
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Lấy danh sách nhân viên thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Lấy danh sách nhân viên thành công"
                  data:
                    type: object
                    properties:
                      staff:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
              example:
                success: true
                message: "Lấy danh sách nhân viên thành công"
                data:
                  staff:
                    - _id: "64f1a2b3c4d5e6f7a8b9c0d5"
                      name: "Trần Thị Bảo"
                      email: "tranthibao@laundry.vn"
                      phone: "0987654321"
                      role: "staff"
                      isActive: true
                    - _id: "64f1a2b3c4d5e6f7a8b9c0d8"
                      name: "Lê Văn Cường"
                      email: "levanncuong@laundry.vn"
                      phone: "0911222333"
                      role: "staff"
                      isActive: true
                pagination:
                  total: 8
                  page: 1
                  limit: 20
                  totalPages: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'